Revert "move new shoto form component into the components folder"
This reverts commit 099172ff4811097b6572a55149e738da1c072da4.
